2 Textures updated. onExit and onEnter events added for Tile
Created good textures for:
  - Wall
  - Floor

Entity movement update:
when an entity enters/exits a tile, a Tile event is triggered.

Player attack added.
Added some code that make doors close when they are passed. However it has a bug that make it close when you go back, so you are stuck. It was commented.